App lock: use keychain directly
