+++ Wire-iOS/Sources/Components/Settings/SettingsPropertyFactory.swift
@@ -65,6 +65,7 @@ protocol CrashlogManager {
 
 extension BITHockeyManager: CrashlogManager {}
 
+
 class SettingsPropertyFactory {
     let userDefaults: UserDefaults
     var analytics: AnalyticsInterface?
@@ -247,18 +248,12 @@ class SettingsPropertyFactory {
             return SettingsBlockProperty(
                 propertyName: propertyName,
                 getAction: { _ in
-                    guard let data = ZMKeychain.data(forAccount: SettingsPropertyName.lockApp.rawValue),
-                            data.count != 0 else {
-                        return SettingsPropertyValue(false)
-                    }
-                    
-                    return SettingsPropertyValue(String(data: data, encoding: .utf8) == "YES")
+                    return SettingsPropertyValue(AppLock.isActive)
             },
                 setAction: { _, value in
                     switch value {
                     case .number(value: let lockApp):
-                        let data = (lockApp.boolValue ? "YES" : "NO").data(using: .utf8)!
-                        ZMKeychain.setData(data, forAccount: SettingsPropertyName.lockApp.rawValue)
+                        AppLock.isActive = lockApp.boolValue
                     default: throw SettingsPropertyError.WrongValue("Incorrect type \(value) for key \(propertyName)")
                     }
             })
@@ -266,28 +261,12 @@ class SettingsPropertyFactory {
             return SettingsBlockProperty(
                 propertyName: propertyName,
                 getAction: { _ in
-                    guard let data = ZMKeychain.data(forAccount: SettingsPropertyName.lockAppLastDate.rawValue),
-                        data.count != 0 else {
-                            return SettingsPropertyValue(0)
-                    }
-                    
-                    let intBits = data.withUnsafeBytes({(bytePointer: UnsafePointer<UInt8>) -> UInt32 in
-                        bytePointer.withMemoryRebound(to: UInt32.self, capacity: 4) { pointer in
-                            return pointer.pointee
-                        }
-                    })
-                    
-                    return SettingsPropertyValue(UInt32(littleEndian: intBits))
+                    return SettingsPropertyValue(AppLock.lastUnlockDateAsInt)
             },
                 setAction: { _, value in
                     switch value {
                     case .number(value: let lockAppLastDate):
-                        var value: UInt32 = lockAppLastDate.uint32Value
-                        let data = withUnsafePointer(to: &value) {
-                            Data(bytes: UnsafePointer($0), count: MemoryLayout.size(ofValue: lockAppLastDate))
-                        }
-                        
-                        ZMKeychain.setData(data, forAccount: SettingsPropertyName.lockAppLastDate.rawValue)
+                        AppLock.lastUnlockDateAsInt = lockAppLastDate.uint32Value
                     default: throw SettingsPropertyError.WrongValue("Incorrect type \(value) for key \(propertyName)")
                     }
             })
